Hectoparsec

Description

Top-level module for Hectoparsec. These parsers are smart in keeping track of labels for parsers, which can be used to generate informational error messages. They also keep track of source positions, which can be attached to results.

To get started, define a parser type for your specific usecase. A parser works over a certain input stream and can use a custom error type or label type. Custom input streams can be used by implementing the Stream typeclass.

import Control.Applicative
import Data.Text (Text)
import Data.Void (Void)
import Hectoparsec

-- Parser over a text stream, with no custom errors, and with string parser labels.
type P = Parser Text Void String

data Color = Red | Green | Blue

red, green, blue :: P Color

red   = Red   <$ string "red"   <?> "red"
green = Green <$ string "green" <?> "green"
blue  = Blue  <$ string "blue"  <?> "blue"

color :: P Color
color = red <|> green <|> blue

parseColor :: FilePath -> Text -> Either (ParseError Text Void String) Color
parseColor fp s = evalParser (color <* endOfInput) fp s

Re-exports

module Hectoparsec.Error

module Hectoparsec.Pos

module Hectoparsec.Stream

module Hectoparsec.State

ParserT monad transformer

data ParserT s e l m a

The type of a parser for a stream s, using custom errors e and custom labels l.

If custom errors or custom labels are not needed, you can simply set it to Void to ignore it. Generally, if your parser cannot error, you would do so. Otherwise, you should set the error and label types to something that would allow you to create useful error messages. In particular, labels are tracked in order to create a list of expected items whenever parsers fail.

ParserT implements MonadParser for the primitive parser combinators, Alternative for branching parsers, and the usual stack of Functor, Applicative, and Monad, along with the classes from mtl.

evalParserT :: Monad m => ParserT s e l m a -> FilePath -> s -> m (Either (ParseError s e l) a)

Runs a parser given an input stream and the file name. Returns either the parse error or the result.

runParserT :: Monad m => ParserT s e l m a -> State s -> m (State s, Either (ParseError s e l) a)

A variant of evalParserT that takes in an initial state and also gives the final state.

Parser monad

type Parser s e l = ParserT s e l Identity

The ParserT type specialized to the Identity monad.

evalParser :: Parser s e l a -> FilePath -> s -> Either (ParseError s e l) a

Runs a parser given an input stream and the file name. Returns either the parse error or the result.

runParser :: Parser s e l a -> State s -> (State s, Either (ParseError s e l) a)

A variant of evalParser that takes in an initial state and also gives the final state.

MonadParser typeclass

class (Stream s, MonadPlus m) => MonadParser s e l m | m -> s e l where

Monad m that implements the primitive parsers for a stream s, using custom errors e and custom labels l. These parsers should have a notion of whether the input was consumed or not. They should also track the parser state, errors, and labels.

The MonadPlus instance should be equal to the Alternative instance, and it should implement the operations for branching parsers. In particular, p <|> q must commit to the first branch that consumes input.

The ParserT instance is the canonical instance for this class.

Methods

matchToken

Arguments

:: (Maybe (Token s) -> Either (ErrorItem s e l) a)
A function to match on tokens. It can return either an error item or the resulting value.

-> m a

Match on a token, returning either the value or an error. If this succeeds, input is consumed.

For matching by equality, use the derived char combinator.

matchTokens

Arguments

:: Int
Length n of the chunk to take. If n <= 0, no tokens are taken.

-> (Chunk s -> Either (ErrorItem s e l) a)
A function to match on the chunk. If the chunk is empty, then n <= 0 or we are at the end.

-> m a

Match on a chunk of at most n length, returning either the value or an error. If it fails, the parser will backtrack the stream. If this succeeds and chunk is non-empty, input is consumed.

For matching by equality, use the derived string combinator.

matchTokenWhile

Arguments

:: (Token s -> Bool)
The predicate to check a token.

-> (Chunk s -> Either (ErrorItem s e l) a)
A function to match on the chunk.

-> m a

Take tokens that satisfy a predicate, and match on them, returning either the value or an error. If it fails, the parser will backtrack the stream. If this succeeds and the chunk is non-empty, input is consumed.

For matching just by a predicate, use the derived tokenWhile and tokenWhile1 combinators.

endOfInput :: m ()

A parser that only succeeds at the end of the stream.

withLabel :: Maybe l -> m a -> m a

Adds or removes a label for a parser. See label and hidden for more information.

By default, no parsers defined in this library are labelled. It is entirely up to you to label parsers.

try :: m a -> m a

Backtracks a parser if it failed. That is, if a parser p fails, then try p will be considered to not have consumed input. This can be used for arbitrary lookahead.

In the example below, alt1 will not act as expected, since red will consume the 'r', meaning rad will not be tried. Adding try in alt2 will allow it to work as expected.

red = char 'r' >> char 'e' >> char 'd'
rad = char 'r' >> char 'a' >> char 'd'
alt1 = red <|> rad
alt2 = try red <|> rad

lookahead :: m a -> m a

Backtracks a parser if it succeeds. That is, if a parser p succeeds, then lookahead p will be considered to not have consumed input.

This does not affect the parser if it fails, i.e. failed parsers can still consume input. Use try along with this function if you need to backtrack on failure too.

notFollowedBy :: m a -> m ()

Creates a parser that only succeeds if the original fails.

This parser never consumes input nor modifies parser state.

recover :: (ParseError s e l -> m a) -> m a -> m a

Creates a parser that can recover from parse failures.

If the recovery parser fails, it will act as if only the original parser failed.

observing :: m a -> m (Either (ParseError s e l) a)

Observes the result of a parser, allowing parsing to continue on failure.

Note that this does not backtrack the parser whether it succeeds or fails.

parseError :: ParseError s e l -> m a

Fails parsing with a parse error.

getState :: m (State s)

Gets the parser state.

putState :: State s -> m ()

Replaces the parser state.

Derived combinators

Input consumption

anyToken :: MonadParser s e l m => m (Token s)

Parses any token.

char :: (MonadParser s e l m, Eq (Token s)) => Token s -> m (Token s)

Parses a specific token. Note that this parser is not labelled by default.

semicolon = char ';'

string :: forall s e l m. (MonadParser s e l m, Eq (Chunk s)) => Chunk s -> m (Chunk s)

Parses a specific sequence of tokens. This fully backtracks, since it uses matchTokens. Note that this parser is not labelled by default.

color = string "red" <|> string "green" <|> string "blue"

satisfy :: MonadParser s e l m => (Token s -> Bool) -> m (Token s)

Parses a token that satisfies a predicate.

digit = satisfy isDigit

peek :: MonadParser s e l m => m (Maybe (Token s))

Peeks at the next token, without advancing the stream in any way.

peekNext :: MonadParser s e l m => m (Token s)

Peeks at the next token, without advancing the stream in any way. If the stream is empty (i.e. there is no next token), an unexpected end of input error is reported.

countTokens :: forall s e l m. MonadParser s e l m => Int -> m (Chunk s)

Parses a chunk of length exactly n, not more, not less. This fully backtracks, since it uses matchTokens.

tokenWhile :: MonadParser s e l m => (Token s -> Bool) -> m (Chunk s)

Takes zero or more tokens that match a predicate. The resulting parser cannot fail. This fully backtracks, since it uses matchTokenWhile. This should be more performant than using many and satisfy.

digits = tokenWhile isDigit

tokenWhile1 :: forall s e l m. MonadParser s e l m => (Token s -> Bool) -> m (Chunk s)

Takes one or more tokens that match a predicate. This fully backtracks, since it uses matchTokenWhile. This should be more performant than using some and satisfy.

digits1 = tokenWhile1 isDigit

matchRest :: MonadParser s e l m => m (Chunk s)

Consumes the rest of the input. This parser cannot fail, though the chunk may be empty.

atEnd :: MonadParser s e l m => m Bool

A parser that checks whether we are at the end of the stream.

Label combinators

label :: MonadParser s e l m => l -> m a -> m a

Adds a label to a parser. This is used for labelling parsers that do not have one for better error messages, or for labelling a complex combination of parsers where you want to give it a more general label instead of merging the labels of each constituent parser.

label lbl p = withLabel (Just lbl) p

(<?>) :: MonadParser s e l m => m a -> l -> m a infix 0

Adds a label to a parser. Simply a synonym for flip label.

hidden :: MonadParser s e l m => m a -> m a

Removes the label from a parser. This can be used to hide labels from errors.

hidden p = withLabel Nothing p

Error combinators

restore :: MonadParser s e l m => (ParseError s e l -> Bool) -> m a -> m a

Restores the state to before using the parser if the error passes a predicate.

The result parser still fails if the given parser fails.

unexpected :: MonadParser s e l m => Unexpected s -> [l] -> m a

Fails parsing with an unexpected item and a list of expected items.

failure :: MonadParser s e l m => String -> m a

Fails parsing with a failure message. These errors are generally for broken invariants.

customError :: MonadParser s e l m => e -> m a

Fails parsing with a custom error.

State combinators

getsState :: MonadParser s e l m => (State s -> a) -> m a

Gets the parser state applied to a function.

getsState f = f <$> getState

modifyState :: MonadParser s e l m => (State s -> State s) -> m ()

Modifies the parser state by a function.

modifyState f = getState >>= putState . f

getInput :: MonadParser s e l m => m s

Gets the input.

getsInput :: MonadParser s e l m => (s -> a) -> m a

Gets the input applied to a function.

getsInput f = f <$> getInput

putInput :: MonadParser s e l m => s -> m ()

Replaces the input.

modifyInput :: MonadParser s e l m => (s -> s) -> m ()

Modifies the input by a function.

modifyInput f = getInput >>= putInput . f

getPos :: MonadParser s e l m => m Pos

Gets the position in the source text.

getOffset :: MonadParser s e l m => m Int

Gets the offset in the input stream.