Hectoparsec.Class

Description

Typeclass for a monad that implements the basic combinators for parsing.

Combinators that are derived from the basic combinators are also defined here.

MonadParser typeclass

class (Stream s, MonadPlus m) => MonadParser s e l m | m -> s e l where

Monad m that implements the primitive parsers for a stream s, using custom errors e and custom labels l.

Methods

matchToken

Arguments

:: (Maybe (Token s) -> Either (ErrorItem s e l) a)
A function to match on tokens. It can return either an error item or the resulting value.

-> m a

Match on a token, returning either the value or an error.

For matching by equality, use the derived char combinator.

matchTokens

Arguments

:: Int
Length n of the chunk to take. If n <= 0, no tokens are taken.

-> (Chunk s -> Either (ErrorItem s e l) a)
A function to match on the chunk. If the chunk is empty, then n <= 0 or we are at the end.

-> m a

Match on a chunk of at most n length, returning either the value or an error. If it fails, the parser will backtrack the stream.

For matching by equality, use the derived string combinator.

matchTokenWhile

Arguments

:: (Token s -> Bool)
The predicate to check a token.

-> (Chunk s -> Either (ErrorItem s e l) a)
A function to match on the chunk.

-> m a

Take tokens that satisfy a predicate, and match on them, returning either the value or an error. If it fails, the parser will backtrack the stream.

For matching just by a predicate, use the derived tokenWhile and tokenWhile1 combinators.

endOfInput :: m ()

A parser that only succeeds at the end of the stream.

withLabel :: Maybe l -> m a -> m a

Adds or removes a label for a parser. See label and hidden for more information.

By default, no parsers defined in this library are labelled. It is entirely up to you to label parsers.

try :: m a -> m a

Backtracks a parser if it failed. This can be used for arbitrary lookahead.

In the example below, alt1 will not act as expected, since red will consume the 'r', meaning rad will not be tried.

red = char 'r' >> char 'e' >> char 'd'
rad = char 'r' >> char 'a' >> char 'd'
alt1 = red <|> rad
alt2 = try red <|> rad

lookahead :: m a -> m a

Backtracks a parser if it succeeds.

This does not affect the parser if it fails, i.e. failed parsers can still consume input. Use try with lookahead if you need to backtrack on failure too.

notFollowedBy :: m a -> m ()

Creates a parser that only succeeds if the original fails.

This parser never consumes input nor modifies parser state.

recover :: (ParseError s e l -> m a) -> m a -> m a

Creates a parser that can recover from parse failures.

If the recovery parser fails, it will act as if only the original parser failed.

observing :: m a -> m (Either (ParseError s e l) a)

Observes the result of a parser, allowing parsing to continue on failure.

Note that this does not backtrack the parser whether it succeeds or fails.

parseError :: ParseError s e l -> m a

Fails parsing with a parse error.

getState :: m (State s)

Gets the parser state.

putState :: State s -> m ()

Replaces the parser state.

Derived combinators

Input consumption

anyToken :: MonadParser s e l m => m (Token s)

Parses any token.

char :: (MonadParser s e l m, Eq (Token s)) => Token s -> m (Token s)

Parses a specific token.

semicolon = char ';'

string :: forall s e l m. (MonadParser s e l m, Eq (Chunk s)) => Chunk s -> m (Chunk s)

Parses a specific sequence of tokens. This fully backtracks, since it uses matchTokens.

color = string "red" <|> string "green" <|> string "blue"

satisfy :: MonadParser s e l m => (Token s -> Bool) -> m (Token s)

Parses a token that satisfies a predicate.

digit = satisfy isDigit

peek :: MonadParser s e l m => m (Maybe (Token s))

Peeks at the next token.

countTokens :: forall s e l m. MonadParser s e l m => Int -> m (Chunk s)

Parses a chunk of length exactly n, not more, not less. This fully backtracks, since it uses matchTokens.

tokenWhile :: MonadParser s e l m => (Token s -> Bool) -> m (Chunk s)

Takes zero or more tokens that match a predicate. The resulting parser cannot fail. This fully backtracks, since it uses matchTokenWhile. This should be more performant than using many and satisfy.

digits = tokenWhile isDigit

tokenWhile1 :: forall s e l m. MonadParser s e l m => (Token s -> Bool) -> m (Chunk s)

Takes one or more tokens that match a predicate. This fully backtracks, since it uses matchTokenWhile. This should be more performant than using some and satisfy.

digits1 = tokenWhile1 isDigit

matchRest :: MonadParser s e l m => m (Chunk s)

Consumes the rest of the input.

Label combinators

label :: MonadParser s e l m => l -> m a -> m a

Adds a label to a parser. This is used for labelling parsers that do not have one for better error messages, or for labelling a complex combination of parsers where you want to give it a more general label instead of merging the labels of each constituent parser.

label lbl p = withLabel (Just lbl) p

(<?>) :: MonadParser s e l m => m a -> l -> m a infix 0

Adds a label to a parser. Simply a synonym for flip label.

hidden :: MonadParser s e l m => m a -> m a

Removes the label from a parser. This can be used to hide labels from errors.

hidden p = withLabel Nothing p

Error reporting

unexpected :: MonadParser s e l m => Unexpected s -> [l] -> m ()

Fails parsing with an unexpected item and a list of expected items.

failure :: MonadParser s e l m => String -> m ()

Fails parsing with a failure message. These errors are generally for broken invariants.

customError :: MonadParser s e l m => e -> m ()

Fails parsing with a custom error.

General combinators

restore :: MonadParser s e l m => (ParseError s e l -> Bool) -> m a -> m a

Restores the state to before using the parser if the error passes a predicate.

The result parser still fails if the given parser fails.

State combinators

getsState :: MonadParser s e l m => (State s -> a) -> m a

Gets the parser state applied to a function.

getsState f = f <$> getState

modifyState :: MonadParser s e l m => (State s -> State s) -> m ()

Modifies the parser state by a function.

modifyState f = getState >>= putState . f

getInput :: MonadParser s e l m => m s

Gets the input.

getsInput :: MonadParser s e l m => (s -> a) -> m a

Gets the input applied to a function.

getsInput f = f <$> getInput

putInput :: MonadParser s e l m => s -> m ()

Replaces the input.

modifyInput :: MonadParser s e l m => (s -> s) -> m ()

Modifies the input by a function.

modifyInput f = getInput >>= putInput . f

getPos :: MonadParser s e l m => m Pos

Gets the stream offset.

getOffset :: MonadParser s e l m => m Int

Gets the stream offset.