Hectoparsec.Error

Description

Data types for errors reported during parsing.

Hectoparsec itself does not come with error pretty printing, but errors can be formatted while prototyping with the functions from Hectoparsec.Debug. External libraries or your own error pretty printer can be used instead.

Parse error data

data Unexpected s

An unexpected item from the input stream.

Constructors

UnexpectedToken (Token s)
An unexpected token.

UnexpectedChunk (Chunk s)
An unexpected chunk.

UnexpectedEnd
An unexpected end of input.

UnexpectedEmpty
An unexpected empty parser.

data Message e

An error message from the input stream.

Constructors

MessageCustom e

MessageFail String

data ErrorItem s e l

A possible error during parsing for a stream s, with error type e and label type l.

Constructors

ErrorItemLabels (Unexpected s) [l]
An error generated through matching tokens. There is an unexpected item and a list of labels from parsers that make up the expected items. There may be duplicates in the list of expected labels, so it is up to you to deduplicate it if needed.

ErrorItemMessages [Message e]
An error generated by fail and failure, and by custom errors, which can generally happen anywhere. These errors are preferred over the errors generated by matching tokens.

data ParseError s e l

A container for an error during parsing for a stream s, with error type e and label type l.

Constructors

ParseError

Fields

parseErrorPos :: !Pos

The position of the error in the source.

parseErrorOffset :: !Int

The offset of the error in the stream.

parseErrorItem :: ErrorItem s e l

The error item.