Hectoparsec.Pos

Description

Data type for source position.

Positional data

data Pos

The position of a token within the source text.

Constructors

Pos

Fields

posFile :: !FilePath

The source file path.

posLine :: !Int

The line number, starting at 1.

posColumn :: !Int

The column number, starting at 1.

initialPos :: FilePath -> Pos

Creates a Pos for the start of a source text (line 1, column 1) with a given file path.

updatePos

Arguments

:: (a -> Bool)
Whether the token is a new line.

-> a
The token.

-> Pos
The position to update.

-> Pos

Updates position based on a character-like token.

Tokens which are considered newlines will increment the line number by 1 and set the column number to 1. Otherwise, the column number is incremented by one.