Hectoparsec.Primitive

Description

Primitive definitions and operations for parser combinators.

Parser types

newtype ParserT s e l m a

The type of a parser for a stream s, using custom errors e and custom labels l.

If custom errors or custom labels are not needed, you can simply set it to Void to ignore it. Generally, if your parser cannot error, you would do so. Otherwise, you should use String or a custom data type for your labels in order to generate informational error messages.

Constructors

ParserT

Fields

unParserT :: Continuations s e l m a

type Parser s e l = ParserT s e l Identity

The ParserT type specialized to the Identity monad.

type ThenOk s l m a r = a -> [l] -> State s -> m r

Continuation for a successful parse.

type ThenErr s e l m r = ParseError s e l -> State s -> m r

Continuation for a failed parse.

type Continuations s e l m a = forall r. State s -> ThenOk s l m a r -> ThenErr s e l m r -> ThenOk s l m a r -> ThenErr s e l m r -> m r

Continuations for a parser.

data Reply s e l a

The reply after parsing.

Constructors

Reply

Fields

replyState :: !(State s)

Resulting state.

replyConsumed :: Bool

Whether the parser consumed or not.

replyResult :: Either (ParseError s e l) a

Either the error or result value.

makeParserT :: Monad m => (State s -> m (Reply s e l a)) -> ParserT s e l m a

Creates a parser based on a reply.

contParserT :: Monad m => ParserT s e l m a -> State s -> m (Reply s e l a)

Runs a parser and retrieve the reply.