Hectoparsec.State

Description

The parser state.

Parser state

data State s

The parser state.

Constructors

State

Fields

stateInput :: s

The rest of the input to parse.

statePos :: !Pos

The current position in the source.

stateOffset :: !Int

The current offset in the stream.

initialState :: FilePath -> s -> State s

Creates a State at the start of a stream.

makeErrorAt :: State s -> ErrorItem s e l -> ParseError s e l

Creates a parse error from an error item located by some parsing state.