Hectoparsec.Debug

Description

Combinators and functions for debugging ParserT.

Re-exports

module Debug.Trace

Debug combinators

pdbg :: (Stream s, Show (Token s), Show (Chunk s), Show e, Show l, Show a) => String -> ParserT s e l m a -> ParserT s e l m a

Wraps a parser with trace for debugging, showing the parser state before and after the parser.

A parser can either consume or not consume tokens, and it can succeed with a parsed value or fail with an error.

(<??>) :: (Stream s, Show (Token s), Show (Chunk s), Show e, Show l, Show a) => ParserT s e l m a -> String -> ParserT s e l m a infix 0

Infix version of flip pdbg.

Debug printers

parseTest :: (Show (Token s), Show (Chunk s), Show e, Show l, Show a) => Parser s e l a -> s -> IO ()

Runs a parser and prints the output or error to the console.

fmtP :: (Show (Token s), Show (Chunk s), Show e, Show l, Show a) => Either (ParseError s e l) a -> String

Formats the result of running a parser for debugging purposes.

fmtE :: (Show (Token s), Show (Chunk s), Show e, Show l) => ParseError s e l -> String

Formats the errors from a parser for debugging purposes.